Source/reactProject/requirements.txt
Django ==3.2.
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3